+++ src/main/java/com/brand/blockus/datagen/providers/BlockusModelProvider.java
@@ -13,7 +13,7 @@ import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
 import net.minecraft.client.data.*;
-import net.minecraft.client.render.model.json.WeightedUnbakedModel;
+import net.minecraft.client.render.model.json.WeightedVariant;
 import net.minecraft.data.family.BlockFamily;
 import net.minecraft.item.Item;
 import net.minecraft.state.property.Properties;
@@ -648,13 +648,13 @@ public class BlockusModelProvider extends FabricModelProvider {
     public final void registerPost(BlockStateModelGenerator modelGenerator, Block block, Block textureSource) {
         TextureMap textureMap = TextureMap.sideAndEndForTop(textureSource);
         Identifier identifier = BlockusModels.TEMPLATE_POST.upload(block, textureMap, modelGenerator.modelCollector);
-        WeightedUnbakedModel weightedUnbakedModel = createModel(BlockusModels.TEMPLATE_POST_CONNECT.upload(block, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(BlockusModels.TEMPLATE_POST_CONNECT_TOP.upload(block, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createPostBlockState(block, createModel(identifier), weightedUnbakedModel, weightedUnbakedModel2, createModel(Blockus.id("block/chain_connect")), createModel(Blockus.id("block/chain_connect_top"))));
+        WeightedVariant modelVariant = createWeightedVariant(BlockusModels.TEMPLATE_POST_CONNECT.upload(block, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant2 = createWeightedVariant(BlockusModels.TEMPLATE_POST_CONNECT_TOP.upload(block, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createPostBlockState(block, createWeightedVariant(identifier), modelVariant, modelVariant2, createWeightedVariant(Blockus.id("block/chain_connect")), createWeightedVariant(Blockus.id("block/chain_connect_top"))));
         modelGenerator.registerParentedItemModel(block, identifier);
     }
 
-    public static BlockModelDefinitionCreator createPostBlockState(Block postBlock, WeightedUnbakedModel postModelId, WeightedUnbakedModel sideModelId, WeightedUnbakedModel topModelId, WeightedUnbakedModel chainModelId, WeightedUnbakedModel topChainModelId) {
+    public static BlockModelDefinitionCreator createPostBlockState(Block postBlock, WeightedVariant postModelId, WeightedVariant sideModelId, WeightedVariant topModelId, WeightedVariant chainModelId, WeightedVariant topChainModelId) {
         return MultipartBlockModelDefinitionCreator.create(postBlock)
             .with(createMultipartConditionBuilder().put(Properties.AXIS, Direction.Axis.X), postModelId.apply(ROTATE_X_90).apply(ROTATE_Y_90))
             .with(createMultipartConditionBuilder().put(Properties.AXIS, Direction.Axis.Y), postModelId)
@@ -675,42 +675,42 @@ public class BlockusModelProvider extends FabricModelProvider {
 
     public final void registerCookieBlock(BlockStateModelGenerator modelGenerator) {
         Block block = BlockusBlocks.COOKIE_BLOCK;
-        WeightedUnbakedModel weightedUnbakedModel = createModel(TexturedModel.CUBE_ALL.upload(block, modelGenerator.modelCollector));
+        WeightedVariant modelVariant = createWeightedVariant(TexturedModel.CUBE_ALL.upload(block, modelGenerator.modelCollector));
         modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block).with(BlockStateVariantMap.models(com.brand.blockus.utils.Properties.BITES_9)
-            .register(0, createModel(ModelIds.getBlockModelId(block)))
-            .register(1, createModel(ModelIds.getBlockSubModelId(block, "_slice1")))
-            .register(2, createModel(ModelIds.getBlockSubModelId(block, "_slice2")))
-            .register(3, createModel(ModelIds.getBlockSubModelId(block, "_slice3")))
-            .register(4, createModel(ModelIds.getBlockSubModelId(block, "_slice4")))
-            .register(5, createModel(ModelIds.getBlockSubModelId(block, "_slice5")))
-            .register(6, createModel(ModelIds.getBlockSubModelId(block, "_slice6")))
-            .register(7, createModel(ModelIds.getBlockSubModelId(block, "_slice7")))
-            .register(8, createModel(ModelIds.getBlockSubModelId(block, "_slice8")))));
+            .register(0, createWeightedVariant(ModelIds.getBlockModelId(block)))
+            .register(1, createWeightedVariant(ModelIds.getBlockSubModelId(block, "_slice1")))
+            .register(2, createWeightedVariant(ModelIds.getBlockSubModelId(block, "_slice2")))
+            .register(3, createWeightedVariant(ModelIds.getBlockSubModelId(block, "_slice3")))
+            .register(4, createWeightedVariant(ModelIds.getBlockSubModelId(block, "_slice4")))
+            .register(5, createWeightedVariant(ModelIds.getBlockSubModelId(block, "_slice5")))
+            .register(6, createWeightedVariant(ModelIds.getBlockSubModelId(block, "_slice6")))
+            .register(7, createWeightedVariant(ModelIds.getBlockSubModelId(block, "_slice7")))
+            .register(8, createWeightedVariant(ModelIds.getBlockSubModelId(block, "_slice8")))));
     }
 
     public final void registerAmethystLamp(BlockStateModelGenerator modelGenerator) {
         Block block = BlockusBlocks.AMETHYST_LAMP;
-        WeightedUnbakedModel weightedUnbakedModel = createModel(TexturedModel.CUBE_ALL.upload(block, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(modelGenerator.createSubModel(block, "_low", Models.CUBE_ALL, TextureMap::all));
-        WeightedUnbakedModel weightedUnbakedModel3 = createModel(modelGenerator.createSubModel(block, "_medium", Models.CUBE_ALL, TextureMap::all));
-        WeightedUnbakedModel weightedUnbakedModel4 = createModel(modelGenerator.createSubModel(block, "_high", Models.CUBE_ALL, TextureMap::all));
+        WeightedVariant modelVariant = createWeightedVariant(TexturedModel.CUBE_ALL.upload(block, modelGenerator.modelCollector));
+        WeightedVariant modelVariant2 = createWeightedVariant(modelGenerator.createSubModel(block, "_low", Models.CUBE_ALL, TextureMap::all));
+        WeightedVariant modelVariant3 = createWeightedVariant(modelGenerator.createSubModel(block, "_medium", Models.CUBE_ALL, TextureMap::all));
+        WeightedVariant modelVariant4 = createWeightedVariant(modelGenerator.createSubModel(block, "_high", Models.CUBE_ALL, TextureMap::all));
         modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block).with(BlockStateVariantMap.models(Properties.POWER)
-            .register(0, createModel(ModelIds.getBlockModelId(block)))
-            .register(1, weightedUnbakedModel2)
-            .register(2, weightedUnbakedModel2)
-            .register(3, weightedUnbakedModel2)
-            .register(4, weightedUnbakedModel2)
-            .register(5, weightedUnbakedModel2)
-            .register(6, weightedUnbakedModel3)
-            .register(7, weightedUnbakedModel3)
-            .register(8, weightedUnbakedModel3)
-            .register(9, weightedUnbakedModel3)
-            .register(10, weightedUnbakedModel3)
-            .register(11, weightedUnbakedModel4)
-            .register(12, weightedUnbakedModel4)
-            .register(13, weightedUnbakedModel4)
-            .register(14, weightedUnbakedModel4)
-            .register(15, weightedUnbakedModel4)));
+            .register(0, createWeightedVariant(ModelIds.getBlockModelId(block)))
+            .register(1, modelVariant2)
+            .register(2, modelVariant2)
+            .register(3, modelVariant2)
+            .register(4, modelVariant2)
+            .register(5, modelVariant2)
+            .register(6, modelVariant3)
+            .register(7, modelVariant3)
+            .register(8, modelVariant3)
+            .register(9, modelVariant3)
+            .register(10, modelVariant3)
+            .register(11, modelVariant4)
+            .register(12, modelVariant4)
+            .register(13, modelVariant4)
+            .register(14, modelVariant4)
+            .register(15, modelVariant4)));
     }
 
     public final void registerSmallHedge(BlockStateModelGenerator modelGenerator, Block hedgeBlock, Block textureSource) {
@@ -723,10 +723,10 @@ public class BlockusModelProvider extends FabricModelProvider {
 
     private void registerSmallHedge(BlockStateModelGenerator modelGenerator, Block hedgeBlock, Block textureSource, boolean isTinted, int tintColor) {
         TextureMap textureMap = TextureMap.of(BlockusTextureKey.HEDGE, TextureMap.getId(textureSource));
-        WeightedUnbakedModel weightedUnbakedModel = createModel(BlockusModels.TEMPLATE_SMALL_HEDGE_END.upload(hedgeBlock, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(BlockusModels.TEMPLATE_SMALL_HEDGE_SIDE.upload(hedgeBlock, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel3 = createModel(BlockusModels.TEMPLATE_SMALL_HEDGE_SIDE_TALL.upload(hedgeBlock, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createWallBlockState(hedgeBlock, weightedUnbakedModel, weightedUnbakedModel2, weightedUnbakedModel3));
+        WeightedVariant modelVariant = createWeightedVariant(BlockusModels.TEMPLATE_SMALL_HEDGE_END.upload(hedgeBlock, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant2 = createWeightedVariant(BlockusModels.TEMPLATE_SMALL_HEDGE_SIDE.upload(hedgeBlock, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant3 = createWeightedVariant(BlockusModels.TEMPLATE_SMALL_HEDGE_SIDE_TALL.upload(hedgeBlock, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createWallBlockState(hedgeBlock, modelVariant, modelVariant2, modelVariant3));
         Identifier identifier = BlockusModels.TEMPLATE_SMALL_HEDGE_INVENTORY.upload(hedgeBlock, textureMap, modelGenerator.modelCollector);
         if (isTinted) {
             modelGenerator.registerTintedItemModel(hedgeBlock, identifier, ItemModels.constantTintSource(tintColor));
@@ -736,13 +736,13 @@ public class BlockusModelProvider extends FabricModelProvider {
     }
 
     public final void registerCarpet(BlockStateModelGenerator modelGenerator, Block wool, Block carpet) {
-        WeightedUnbakedModel weightedUnbakedModel = createModel(TexturedModel.CARPET.get(wool).upload(carpet, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createSingletonBlockState(carpet, weightedUnbakedModel));
+        WeightedVariant modelVariant = createWeightedVariant(TexturedModel.CARPET.get(wool).upload(carpet, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createSingletonBlockState(carpet, modelVariant));
     }
 
     public final void registerDirectionalCarpet(BlockStateModelGenerator modelGenerator, Block block, Block carpet) {
         TexturedModel.CARPET.get(block).upload(carpet, modelGenerator.modelCollector);
-        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(carpet, createModel(ModelIds.getBlockModelId(carpet))).coordinate(createUpDefaultRotationStates()));
+        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(carpet, createWeightedVariant(ModelIds.getBlockModelId(carpet))).coordinate(createUpDefaultRotationStates()));
     }
 
     public final void registerBlockStairsSlabWithBottom(BlockStateModelGenerator modelGenerator, Block block, Block stairs, Block slab, Block base) {
@@ -797,25 +797,25 @@ public class BlockusModelProvider extends FabricModelProvider {
 
     public final void registerButton(BlockStateModelGenerator modelGenerator, Block buttonBlock, Identifier textureSource) {
         TextureMap textureMap = TextureMap.texture(textureSource);
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.BUTTON.upload(buttonBlock, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(Models.BUTTON_PRESSED.upload(buttonBlock, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant = createWeightedVariant(Models.BUTTON.upload(buttonBlock, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant2 = createWeightedVariant(Models.BUTTON_PRESSED.upload(buttonBlock, textureMap, modelGenerator.modelCollector));
         Identifier identifier = Models.BUTTON_INVENTORY.upload(buttonBlock, textureMap, modelGenerator.modelCollector);
-        modelGenerator.blockStateCollector.accept(createButtonBlockState(buttonBlock, weightedUnbakedModel, weightedUnbakedModel2));
+        modelGenerator.blockStateCollector.accept(createButtonBlockState(buttonBlock, modelVariant, modelVariant2));
         modelGenerator.registerParentedItemModel(buttonBlock, identifier);
     }
 
     public final void registerPressurePlate(BlockStateModelGenerator modelGenerator, Block pressurePlateBlock, Identifier textureSource) {
         TextureMap textureMap = TextureMap.texture(textureSource);
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.PRESSURE_PLATE_UP.upload(pressurePlateBlock, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(Models.PRESSURE_PLATE_DOWN.upload(pressurePlateBlock, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createPressurePlateBlockState(pressurePlateBlock, weightedUnbakedModel, weightedUnbakedModel2));
+        WeightedVariant modelVariant = createWeightedVariant(Models.PRESSURE_PLATE_UP.upload(pressurePlateBlock, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant2 = createWeightedVariant(Models.PRESSURE_PLATE_DOWN.upload(pressurePlateBlock, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createPressurePlateBlockState(pressurePlateBlock, modelVariant, modelVariant2));
     }
 
     public final void registerRedstoneLamp(BlockStateModelGenerator modelGenerator, Block block, Block lit) {
         registerLitRedstoneLamp(modelGenerator, block, lit);
-        WeightedUnbakedModel weightedUnbakedModel = createModel(TexturedModel.CUBE_ALL.upload(block, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(modelGenerator.createSubModel(block, "_on", Models.CUBE_ALL, TextureMap::all));
-        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block).with(createBooleanModelMap(Properties.LIT, weightedUnbakedModel2, weightedUnbakedModel)));
+        WeightedVariant modelVariant = createWeightedVariant(TexturedModel.CUBE_ALL.upload(block, modelGenerator.modelCollector));
+        WeightedVariant modelVariant2 = createWeightedVariant(modelGenerator.createSubModel(block, "_on", Models.CUBE_ALL, TextureMap::all));
+        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block).with(createBooleanModelMap(Properties.LIT, modelVariant2, modelVariant)));
 
     }
 
@@ -825,9 +825,9 @@ public class BlockusModelProvider extends FabricModelProvider {
     }
 
     public void createAxisRotatedCubeColumn(BlockStateModelGenerator modelGenerator, Block block, TextureMap textureMap) {
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.CUBE_COLUMN.upload(block, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(Models.CUBE_COLUMN_HORIZONTAL.upload(block, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createAxisRotatedBlockState(block, weightedUnbakedModel, weightedUnbakedModel2));
+        WeightedVariant modelVariant = createWeightedVariant(Models.CUBE_COLUMN.upload(block, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant2 = createWeightedVariant(Models.CUBE_COLUMN_HORIZONTAL.upload(block, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createAxisRotatedBlockState(block, modelVariant, modelVariant2));
     }
 
     public final void registerAxisRotatedCubeColumn(BlockStateModelGenerator modelGenerator, Block block) {
@@ -861,14 +861,14 @@ public class BlockusModelProvider extends FabricModelProvider {
     }
 
     public final void registerUpDefaultFacingBlock(BlockStateModelGenerator modelGenerator, Block block) {
-        WeightedUnbakedModel weightedUnbakedModel = createModel(TexturedModel.CUBE_ALL.upload(block, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block, weightedUnbakedModel).coordinate(createUpDefaultRotationStates()));
+        WeightedVariant modelVariant = createWeightedVariant(TexturedModel.CUBE_ALL.upload(block, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block, modelVariant).coordinate(createUpDefaultRotationStates()));
     }
 
     public final void registerTopBottomFacingBottom(BlockStateModelGenerator modelGenerator, Block block) {
         TextureMap textureMap = sideTopBottom(TextureMap.getId(block), TextureMap.getSubId(block, "_top"), TextureMap.getSubId(block, "_bottom"));
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.CUBE_BOTTOM_TOP.upload(block, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block, weightedUnbakedModel).coordinate(createUpDefaultRotationStates()));
+        WeightedVariant modelVariant = createWeightedVariant(Models.CUBE_BOTTOM_TOP.upload(block, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block, modelVariant).coordinate(createUpDefaultRotationStates()));
     }
 
     public final void registerPumpkins(BlockStateModelGenerator modelGenerator, Block block) {
@@ -903,8 +903,8 @@ public class BlockusModelProvider extends FabricModelProvider {
 
     public final void registerBreadBox(BlockStateModelGenerator modelGenerator, Block block) {
         TextureMap textureMap = frontTopSideBottom(block);
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.ORIENTABLE_WITH_BOTTOM.upload(block, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block, weightedUnbakedModel).coordinate(BlockStateVariantMap.operations(Properties.HORIZONTAL_FACING).register(Direction.EAST, ROTATE_Y_90).register(Direction.SOUTH, ROTATE_Y_180).register(Direction.WEST, ROTATE_Y_270).register(Direction.NORTH, NO_OP)));
+        WeightedVariant modelVariant = createWeightedVariant(Models.ORIENTABLE_WITH_BOTTOM.upload(block, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block, modelVariant).coordinate(BlockStateVariantMap.operations(Properties.HORIZONTAL_FACING).register(Direction.EAST, ROTATE_Y_90).register(Direction.SOUTH, ROTATE_Y_180).register(Direction.WEST, ROTATE_Y_270).register(Direction.NORTH, NO_OP)));
     }
 
     public void registerLegacyStonecutter(BlockStateModelGenerator modelGenerator, Block block) {
@@ -913,29 +913,29 @@ public class BlockusModelProvider extends FabricModelProvider {
     }
 
     private void registerDiagonalTimberFrame(BlockStateModelGenerator modelGenerator, Block block) {
-        WeightedUnbakedModel weightedUnbakedModel = createModel(ModelIds.getBlockModelId(block));
+        WeightedVariant modelVariant = createWeightedVariant(ModelIds.getBlockModelId(block));
         Identifier identifier = TextureMap.getSubId(block, "_right");
         Identifier identifier2 = TextureMap.getSubId(block, "_left");
         TextureMap textureMap = (new TextureMap()).put(TextureKey.PARTICLE, identifier).put(TextureKey.NORTH, identifier).put(TextureKey.SOUTH, identifier).put(TextureKey.EAST, identifier).put(TextureKey.WEST, identifier2).put(TextureKey.DOWN, identifier2).put(TextureKey.UP, identifier2);
         Models.CUBE.upload(block, textureMap, modelGenerator.modelCollector);
         modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block).with(BlockStateVariantMap.models(Properties.HORIZONTAL_FACING)
-            .register(Direction.NORTH, weightedUnbakedModel)
-            .register(Direction.SOUTH, weightedUnbakedModel.apply(ROTATE_Y_180))
-            .register(Direction.WEST, weightedUnbakedModel.apply(ROTATE_Y_270))
-            .register(Direction.EAST, weightedUnbakedModel.apply(ROTATE_Y_90))));
+            .register(Direction.NORTH, modelVariant)
+            .register(Direction.SOUTH, modelVariant.apply(ROTATE_Y_180))
+            .register(Direction.WEST, modelVariant.apply(ROTATE_Y_270))
+            .register(Direction.EAST, modelVariant.apply(ROTATE_Y_90))));
     }
 
     public final void registerBeveledGlassPane(BlockStateModelGenerator modelGenerator, Block glass, Block glassPane) {
         modelGenerator.registerSimpleCubeAll(glass);
         TextureMap textureMap = TextureMap.paneAndTopForEdge(glass, BlockusBlocks.BEVELED_GLASS_PANE);
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.TEMPLATE_GLASS_PANE_POST.upload(glassPane, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(Models.TEMPLATE_GLASS_PANE_SIDE.upload(glassPane, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel3 = createModel(Models.TEMPLATE_GLASS_PANE_SIDE_ALT.upload(glassPane, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel4 = createModel(Models.TEMPLATE_GLASS_PANE_NOSIDE.upload(glassPane, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel5 = createModel(Models.TEMPLATE_GLASS_PANE_NOSIDE_ALT.upload(glassPane, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant = createWeightedVariant(Models.TEMPLATE_GLASS_PANE_POST.upload(glassPane, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant2 = createWeightedVariant(Models.TEMPLATE_GLASS_PANE_SIDE.upload(glassPane, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant3 = createWeightedVariant(Models.TEMPLATE_GLASS_PANE_SIDE_ALT.upload(glassPane, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant4 = createWeightedVariant(Models.TEMPLATE_GLASS_PANE_NOSIDE.upload(glassPane, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant5 = createWeightedVariant(Models.TEMPLATE_GLASS_PANE_NOSIDE_ALT.upload(glassPane, textureMap, modelGenerator.modelCollector));
         Item item = glassPane.asItem();
         modelGenerator.registerItemModel(item, modelGenerator.uploadBlockItemModel(item, glass));
-        modelGenerator.blockStateCollector.accept(MultipartBlockModelDefinitionCreator.create(glassPane).with(weightedUnbakedModel).with(createMultipartConditionBuilder().put(Properties.NORTH, true), weightedUnbakedModel2).with(createMultipartConditionBuilder().put(Properties.EAST, true), weightedUnbakedModel2.apply(ROTATE_Y_90)).with(createMultipartConditionBuilder().put(Properties.SOUTH, true), weightedUnbakedModel3).with(createMultipartConditionBuilder().put(Properties.WEST, true), weightedUnbakedModel3.apply(ROTATE_Y_90)).with(createMultipartConditionBuilder().put(Properties.NORTH, false), weightedUnbakedModel4).with(createMultipartConditionBuilder().put(Properties.EAST, false), weightedUnbakedModel5).with(createMultipartConditionBuilder().put(Properties.SOUTH, false), weightedUnbakedModel5.apply(ROTATE_Y_90)).with(createMultipartConditionBuilder().put(Properties.WEST, false), weightedUnbakedModel4.apply(ROTATE_Y_270)));
+        modelGenerator.blockStateCollector.accept(MultipartBlockModelDefinitionCreator.create(glassPane).with(modelVariant).with(createMultipartConditionBuilder().put(Properties.NORTH, true), modelVariant2).with(createMultipartConditionBuilder().put(Properties.EAST, true), modelVariant2.apply(ROTATE_Y_90)).with(createMultipartConditionBuilder().put(Properties.SOUTH, true), modelVariant3).with(createMultipartConditionBuilder().put(Properties.WEST, true), modelVariant3.apply(ROTATE_Y_90)).with(createMultipartConditionBuilder().put(Properties.NORTH, false), modelVariant4).with(createMultipartConditionBuilder().put(Properties.EAST, false), modelVariant5).with(createMultipartConditionBuilder().put(Properties.SOUTH, false), modelVariant5.apply(ROTATE_Y_90)).with(createMultipartConditionBuilder().put(Properties.WEST, false), modelVariant4.apply(ROTATE_Y_270)));
     }
 
     public static void registerColoredTiles(BlockStateModelGenerator modelGenerator, Block block, Block tile1, Block tile2) {
@@ -943,8 +943,8 @@ public class BlockusModelProvider extends FabricModelProvider {
         BlockusModels.CUBE_TILES.upload(block, textures, modelGenerator.modelCollector);
         BlockusModels.CUBE_TILES_2.upload(block, textures, modelGenerator.modelCollector);
         modelGenerator.blockStateCollector.accept(VariantsBlockModelDefinitionCreator.of(block).with(BlockStateVariantMap.models(Properties.HORIZONTAL_AXIS)
-            .register(Direction.Axis.X, createModel(ModelIds.getBlockModelId(block)))
-            .register(Direction.Axis.Z, createModel(ModelIds.getBlockSubModelId(block, "_2")))));
+            .register(Direction.Axis.X, createWeightedVariant(ModelIds.getBlockModelId(block)))
+            .register(Direction.Axis.Z, createWeightedVariant(ModelIds.getBlockSubModelId(block, "_2")))));
     }
 
     public final void registerColoredTilesSimple(BlockStateModelGenerator modelGenerator, Block block) {
@@ -953,33 +953,33 @@ public class BlockusModelProvider extends FabricModelProvider {
     }
 
     public final void createBlock(BlockStateModelGenerator modelGenerator, Block block, Model model, TextureMap textureMap) {
-        modelGenerator.blockStateCollector.accept(createSingletonBlockState(block, createModel(model.upload(block, textureMap, modelGenerator.modelCollector))));
+        modelGenerator.blockStateCollector.accept(createSingletonBlockState(block, createWeightedVariant(model.upload(block, textureMap, modelGenerator.modelCollector))));
     }
 
     public final void createBlock(BlockStateModelGenerator modelGenerator, Block block, TextureMap textureMap) {
-        modelGenerator.blockStateCollector.accept(createSingletonBlockState(block, createModel(Models.CUBE_ALL.upload(block, textureMap, modelGenerator.modelCollector))));
+        modelGenerator.blockStateCollector.accept(createSingletonBlockState(block, createWeightedVariant(Models.CUBE_ALL.upload(block, textureMap, modelGenerator.modelCollector))));
     }
 
     public final void createStairs(BlockStateModelGenerator modelGenerator, Block block, TextureMap textureMap) {
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.INNER_STAIRS.upload(block, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant = createWeightedVariant(Models.INNER_STAIRS.upload(block, textureMap, modelGenerator.modelCollector));
         Identifier identifier = Models.STAIRS.upload(block, textureMap, modelGenerator.modelCollector);
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(Models.OUTER_STAIRS.upload(block, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createStairsBlockState(block, weightedUnbakedModel, createModel(identifier), weightedUnbakedModel2));
+        WeightedVariant modelVariant2 = createWeightedVariant(Models.OUTER_STAIRS.upload(block, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createStairsBlockState(block, modelVariant, createWeightedVariant(identifier), modelVariant2));
         modelGenerator.registerParentedItemModel(block, identifier);
     }
 
     public final void createSlab(BlockStateModelGenerator modelGenerator, Block block, TextureMap textureMap, Identifier doubleSlab) {
         Identifier identifier = Models.SLAB.upload(block, textureMap, modelGenerator.modelCollector);
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.SLAB_TOP.upload(block, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createSlabBlockState(block, createModel(identifier), weightedUnbakedModel, createModel(doubleSlab)));
+        WeightedVariant modelVariant = createWeightedVariant(Models.SLAB_TOP.upload(block, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createSlabBlockState(block, createWeightedVariant(identifier), modelVariant, createWeightedVariant(doubleSlab)));
         modelGenerator.registerParentedItemModel(block, identifier);
     }
 
     public final void createWall(BlockStateModelGenerator modelGenerator, Block block, TextureMap textureMap) {
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.TEMPLATE_WALL_POST.upload(block, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel2 = createModel(Models.TEMPLATE_WALL_SIDE.upload(block, textureMap, modelGenerator.modelCollector));
-        WeightedUnbakedModel weightedUnbakedModel3 = createModel(Models.TEMPLATE_WALL_SIDE_TALL.upload(block, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createWallBlockState(block, weightedUnbakedModel, weightedUnbakedModel2, weightedUnbakedModel3));
+        WeightedVariant modelVariant = createWeightedVariant(Models.TEMPLATE_WALL_POST.upload(block, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant2 = createWeightedVariant(Models.TEMPLATE_WALL_SIDE.upload(block, textureMap, modelGenerator.modelCollector));
+        WeightedVariant modelVariant3 = createWeightedVariant(Models.TEMPLATE_WALL_SIDE_TALL.upload(block, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createWallBlockState(block, modelVariant, modelVariant2, modelVariant3));
         Identifier identifier = Models.WALL_INVENTORY.upload(block, textureMap, modelGenerator.modelCollector);
         modelGenerator.registerParentedItemModel(block, identifier);
     }
@@ -1043,8 +1043,8 @@ public class BlockusModelProvider extends FabricModelProvider {
     }
 
     public final void registerTintableCrossBlockState(BlockStateModelGenerator modelGenerator, Block block, TextureMap crossTexture) {
-        WeightedUnbakedModel weightedUnbakedModel = createModel(Models.CROSS.upload(block, crossTexture, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createSingletonBlockState(block, weightedUnbakedModel));
+        WeightedVariant modelVariant = createWeightedVariant(Models.CROSS.upload(block, crossTexture, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createSingletonBlockState(block, modelVariant));
     }
 
     public final void registerFlowerPotPlantAndItem(BlockStateModelGenerator modelGenerator, Block block, Block flowerPotBlock, CrossType crossType) {
@@ -1055,7 +1055,7 @@ public class BlockusModelProvider extends FabricModelProvider {
     public final void registerFlowerPotPlant(BlockStateModelGenerator modelGenerator, Block plantBlock, Block flowerPotBlock, CrossType tintType) {
         this.registerTintableCrossBlockState(modelGenerator, plantBlock, tintType);
         TextureMap textureMap = tintType.getFlowerPotTextureMap(plantBlock);
-        WeightedUnbakedModel weightedUnbakedModel = createModel(tintType.getFlowerPotCrossModel().upload(flowerPotBlock, textureMap, modelGenerator.modelCollector));
-        modelGenerator.blockStateCollector.accept(createSingletonBlockState(flowerPotBlock, weightedUnbakedModel));
+        WeightedVariant modelVariant = createWeightedVariant(tintType.getFlowerPotCrossModel().upload(flowerPotBlock, textureMap, modelGenerator.modelCollector));
+        modelGenerator.blockStateCollector.accept(createSingletonBlockState(flowerPotBlock, modelVariant));
     }
 }