+++ changelog.md
@@ -1,5 +1,6 @@
 * Updated to 1.21.5
 * Added fallen white oak in biomes with white oak trees.
-* White oak trees now generate in dark forests.
-* White oak trees are now slightly more common, especially in plains.
 * Leaf litters now generate under every naturally generated white oak trees, even in plains.
+* Wandering Trader now sells 8 White Oak Logs as a special offer.
+* Rainbow Rose now spread using bone meal.
+* Changed Rainbow Petals ID from `rainbow_petal` to `rainbow_petals`